FIX: maxhp, maxmp

+++ src/map/battleentity.cpp
@@ -89,7 +89,7 @@ uint8 CBattleEntity::GetHPP()
 	return (uint8)ceil(((float)health.hp / (float)GetMaxHP()) * 100);
 }
 
-uint32 CBattleEntity::GetMaxHP()
+int32 CBattleEntity::GetMaxHP()
 {
     return (health.maxhp + getMod(MOD_HP)) * (100 + getMod(MOD_HPP)) / 100; //  + (MOD_CONVMPTOHP - MOD_CONVHPTOMP)
 }
@@ -105,7 +105,7 @@ uint8 CBattleEntity::GetMPP()
 	return (uint8)ceil(((float)health.mp / (float)GetMaxMP()) * 100);
 }
 
-uint32 CBattleEntity::GetMaxMP()
+int32 CBattleEntity::GetMaxMP()
 {
     return (health.maxmp + getMod(MOD_MP)) * (100 + getMod(MOD_MPP)) / 100; //  + (MOD_CONVHPTOMP - MOD_CONVMPTOHP)
 }
@@ -130,11 +130,11 @@ uint16 CBattleEntity::addTP(int16 tp)
 *																		*
 ************************************************************************/
 
-uint16 CBattleEntity::addHP(int16 hp)
+int32 CBattleEntity::addHP(int32 hp)
 {
 	if (status == STATUS_NORMAL) status = STATUS_UPDATE;
 
-    int16 cap = cap_value(health.hp + hp, 0, GetMaxHP());
+    int32 cap = cap_value(health.hp + hp, 0, GetMaxHP());
 	hp = health.hp - cap;
 	health.hp = cap;
 
@@ -147,9 +147,9 @@ uint16 CBattleEntity::addHP(int16 hp)
 	return abs(hp);
 }
 
-uint16 CBattleEntity::addMP(int16 mp)
+int32 CBattleEntity::addMP(int32 mp)
 {
-	int16 cap = cap_value(health.mp + mp, 0, GetMaxMP());
+	int32 cap = cap_value(health.mp + mp, 0, GetMaxMP());
 	mp = health.mp - cap;
 	health.mp = cap;
 	return abs(mp);
